Conflicts:
	CHANGELOG.mkd
	script/bootstrap/pip-requirements.txt

+++ src/can/canutil.h
@@ -174,7 +174,7 @@ struct CanBus {
     unsigned int speed;
     short address;
     void* controller;
-    unsigned short maxMessageFrequency;
+    float maxMessageFrequency;
     bool rawWritable;
     void (*interruptHandler)();
     AcceptanceFilterList acceptanceFilters;
